# Conflicts:
#	Source/Blazorise.Bulma/wwwroot/blazorise.bulma.css
#	Source/Blazorise.Bulma/wwwroot/blazorise.bulma.min.css

+++ Source/Extensions/Blazorise.Charts.Streaming/ChartStreamingOptions.cs
@@ -32,7 +32,7 @@ public class ChartStreamingOptions
     /// </summary>
     [JsonPropertyName( "delay" )]
     [JsonIgnore( Condition = JsonIgnoreCondition.WhenWritingNull )]
-    public int? Delay { get; set; } = 0;
+    public long? Delay { get; set; } = 0;
 
     /// <summary>
     /// Frequency at which the chart is drawn on a display (frames per second). This option can be set at chart level but not at axis level. Decrease this value to save CPU power (https://github.com/nagix/chartjs-plugin-streaming#lowering-cpu-usage).
